+++ src/app/resume/page.tsx
@@ -1,4 +1,4 @@
-import { LoopMDX } from "@/feature/MDX/LoopMDX";
+import { LoopMDX } from "@/feature/MDX/components/LoopMDX";
 import Portal from "@/feature/common/components/Portal";
 import RevalidatePost from "@/feature/post/components/RevalidatePost";
 import { REVALIDATE_TAGNAME } from "@/feature/common/constants";
